requirements-hf.txt
gradio>=4.0.0
openai>=1.0.0
python-docx>=0.8.11
requests>=2.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
markdown>=3.4.0
pillow>=9.4.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0